unify ziospec
